fixed to not cast to int (could be > 2gb)

+++ src/suneido/language/builtin/MemoryArena.java
@@ -6,7 +6,7 @@ public class MemoryArena extends SuFunction0 {
 
 	@Override
 	public Object call0() {
-		return (int) Runtime.getRuntime().totalMemory();
+		return Runtime.getRuntime().totalMemory();
 	}
 
 }
